archive/requirements.txt
numpy
biopython
matplotlib
biom-format
ete3
